perf: lazy load motion/react in Toast, Popover and Tooltip
Keep motion/react (framer-motion) out of the static module graph by
lazy-loading the motion-bearing modules and rendering a synchronous,
motion-free path for the non-animated default.

- Popover/Tooltip: introduce static and animated Layer variants; only
  the animated variant imports motion at runtime and re-exports
  AnimatePresence. The parent loads it via React.lazy behind a
  never-resetting latch, wrapped in Suspense.
- Toast: Toast becomes a generic React.lazy + Suspense wrapper over a
  new ToastCard motion impl; ToastProvider gates a lazy ToastList behind
  a hasPushed latch.
- Convert motion types in the public PopoverLayer/TooltipLayer modules
  to type-only imports so the public d.ts is unchanged.
- Add a no-restricted-imports motion guard with a lazy-chunk whitelist.
